Rename files to avoid libtool same member name warning on macOS
Avoid having the same name for cpp files in different directories.
