+++ source/opt/eliminate_dead_constant_pass.h
@@ -15,7 +15,7 @@
 #ifndef LIBSPIRV_OPT_ELIMINATE_DEAD_CONSTANT_PASS_H_
 #define LIBSPIRV_OPT_ELIMINATE_DEAD_CONSTANT_PASS_H_
 
-#include "module.h"
+#include "opt_module.h"
 #include "pass.h"
 
 namespace spvtools {
