+++ source/opt/null_pass.h
@@ -15,7 +15,7 @@
 #ifndef LIBSPIRV_OPT_NULL_PASS_H_
 #define LIBSPIRV_OPT_NULL_PASS_H_
 
-#include "module.h"
+#include "opt_module.h"
 #include "pass.h"
 
 namespace spvtools {
